+++ src/modules/proxy/service.ts
@@ -84,12 +84,15 @@ export abstract class ProxyService {
         const prefix = isTsLink ? this.TS_PREFIX : this.M3U8_PREFIX;
         if (all && trimmedLine.startsWith("http")) {
           // https://yourproxy.com/?url=https://somevideo.m3u8&all=yes
-          return `${this.SCHEMA}${host}${prefix}?url=${trimmedLine}`;
+          return `${this.SCHEMA}${host}${prefix}?url=${btoa(
+            trimmedLine,
+          )}&format=base64`;
         }
 
         // autoencodes all fields
         const params = new URLSearchParams({
-          url: `${targetFilename}${trimmedLine}`,
+          url: btoa(`${targetFilename}${trimmedLine}`),
+          format: "base64",
         });
         if (origin) {
           params.append("origin", origin);